refactor: 핸드폰 - 추가 함수 별도 Utility 폴더로 분리
