+++ ContactManager/ContactManager/Controller/AddContactViewController.swift
@@ -60,29 +60,3 @@ class AddContactViewController: UIViewController {
     sender.text = sender.text?.formatingPhone()
   }
 }
-
-// MARK: - extension
-
-extension String {
-  func formatingPhone() -> String {
-    let _str = self.replacingOccurrences(of: "-", with: "")
-    if(self.count > 12) { return self }
-    
-    var regexPattern: String
-    if _str.count == 9 {
-      regexPattern = "([0-9]{0,2})([0-9]{0,3})([0-9]{0,4})"
-    } else {
-      regexPattern = "([0-9]{0,3})([0-9]{0,4})([0-9]{0,4})"
-    }
-    
-    if let regex = try? NSRegularExpression(pattern: regexPattern, options: .caseInsensitive) {
-      let modString = regex.stringByReplacingMatches(in: _str,
-                                                     options: [],
-                                                     range: NSRange(_str.startIndex..., in: _str),
-                                                     withTemplate: "$1-$2-$3")
-      return modString.trimmingCharacters(in: ["-"])
-    }
-    
-    return _str
-  }
-}
